+++ DFC.App.CareerPath.IntegrationTests.API/Support/Enums/ActionType.cs
@@ -1,0 +1,8 @@
+﻿namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Support.Enums
+{
+    public enum ActionType
+    {
+        Published,
+        Deleted,
+    }
+}
