+++ DFC.App.CareerPath.IntegrationTests.API/Support/Enums/CType.cs
@@ -1,0 +1,12 @@
+﻿namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Support.Enums
+{
+    public enum CType
+    {
+        JobProfile,
+        JobProfileSoc,
+        WorkingHoursDetails,
+        WorkingPattern,
+        WorkingPatternDetails,
+        JobProfileSpecialism,
+    }
+}
