+++ DFC.App.CareerPath.IntegrationTests.API/Test/CareerPathTest.cs
@@ -12,7 +12,7 @@ namespace DFC.App.RelatedCareers.Tests.IntegrationTests.API.Test
         [Description("Tests that the CType 'JobProfile' successfully tiggers a related career path update to an existing job profile")]
         public async Task CareerPathJobProfile()
         {
-            Response<CareerPathAPIResponse> response = await this.CommonAction.ExecuteGetRequest<CareerPathAPIResponse>(Settings.APIConfig.EndpointBaseUrl.Replace("{id}", this.JobProfile.JobProfileId, System.StringComparison.CurrentCultureIgnoreCase), GetRequest.ContentType.Json).ConfigureAwait(true);
+            Response<CareerPathAPIResponse> response = await this.CommonAction.ExecuteGetRequest<CareerPathAPIResponse>(Settings.APIConfig.EndpointBaseUrl.Replace("{id}", this.JobProfile.JobProfileId, System.StringComparison.CurrentCultureIgnoreCase), ContentType.Json).ConfigureAwait(true);
             Assert.AreEqual(this.JobProfile.CareerPathAndProgression, response.Data.CareerPathAndProgression[0]);
         }
     }
