+++ DFC.App.CareerPath.Tests.Common.APISupport/GetRequest.cs
@@ -1,36 +1,33 @@
 ﻿using Newtonsoft.Json;
 using RestSharp;
-using System.Collections.Generic;
 using System.Net;
 using System.Threading;
 
 namespace DFC.App.RelatedCareers.Tests.Common.APISupport
 {
+    public enum ContentType
+    {
+        Json,
+        Html,
+    }
+
     public class GetRequest
     {
-        private RestRequest Request { get; set; }
-        public Dictionary<string, string> Headers { get; } = new Dictionary<string, string>();
-
         public GetRequest(string endpoint)
         {
-            Request = new RestRequest(endpoint, Method.GET);
+            this.Request = new RestRequest(endpoint, Method.GET);
         }
 
-        public void AddQueryParameter(string name, string value)
-        {
-            Request.AddParameter(name, value);
-        }
+        private RestRequest Request { get; set; }
 
         public void AddVersionHeader(string version)
         {
             Request.AddHeader("version", version);
-            Headers.Add("version", version);
         }
 
         public void AddApimKeyHeader(string apimSubscriptionKey)
         {
             Request.AddHeader("Ocp-Apim-Subscription-Key", apimSubscriptionKey);
-            Headers.Add("Ocp-Apim-Subscription-Key", apimSubscriptionKey);
         }
 
         public void AddAcceptHeader(ContentType contentType)
@@ -39,36 +36,12 @@ namespace DFC.App.RelatedCareers.Tests.Common.APISupport
             {
                 case ContentType.Json:
                     Request.AddHeader("Accept", "application/json");
-                    Headers.Add("Accept", "application/json");
                     break;
 
                 case ContentType.Html:
                     Request.AddHeader("Accept", "text/html");
-                    Headers.Add("Accept", "text/html");
                     break;
             }
-        }
-
-        public void AddContentType(ContentType contentType)
-        {
-            switch(contentType)
-            {
-                case ContentType.Json:
-                    Request.AddHeader("Content-Type", "application/json");
-                    Headers.Add("Content-Type", "application/json");
-                    break;
-
-                case ContentType.Html:
-                    Request.AddHeader("Content-Type", "text/html");
-                    Headers.Add("Content-Type", "text/html");
-                    break;
-            }
-        }
-
-        public enum ContentType
-        {
-            Json,
-            Html
         }
 
         public Response<T> Execute<T>()
@@ -76,8 +49,8 @@ namespace DFC.App.RelatedCareers.Tests.Common.APISupport
             AutoResetEvent autoResetEvent = new AutoResetEvent(false);
             Response<T> response = new Response<T>();
             IRestResponse rawResponse = null;
-            
-            new RestClient().ExecuteAsync(Request, (IRestResponse apiResponse) => { 
+
+            new RestClient().ExecuteAsync(Request, (IRestResponse apiResponse) => {
                 rawResponse = apiResponse;
                 autoResetEvent.Set();
             });
